Move cfunctions to .codegen, support NAN, refactor reserved_words
